Move sketch to appropriately named subfolder
Arduino IDE requires that a sketch be located in a folder with name matching that of the primary `.ino` file.

When opening a file that does not meet this folder name matching requirement, Arduino IDE presents a seemingly helpful
dialog:

> The file "SigGen.ino" needs to be inside a sketch folder named "SigGen".
> Create this folder, move the file, and continue?

When the user clicks the "OK" button in the dialog, Arduino IDE moves only the file "SigGen.ino" to the new folder,
leaving behind the other source files. This causes compilation of the sketch to fail:

```
Settings_Module.h: No such file or directory
```

The alternative solution that might come to mind is to change the sketch name to match the repository name. This would
work for people who use Git to clone the repository, but when GitHub's popular source code archive download feature
(e.g., "Clone or download > Download ZIP") is used to download the contents of a repository the Git ref
(branch/tag/commit) name is appended to the folder name, resulting in the sketch and folder names still not matching.